Reorganized files. Added linter files. Moved window listener to app.js. #1
