+++ .eslintrc
@@ -1,0 +1,32 @@
+{
+  "env": {
+    "browser": true,
+    "node": true,
+    "es6": true
+  },
+  "extends": ["standard", "plugin:react/recommended"],
+  "rules": {
+    "padded-blocks": 0,
+    "curly": ["error", "multi-line"],
+    "brace-style": ["error", "stroustrup"],
+    "react/prop-types": 0,
+    "space-before-function-paren": ["error", {
+      "anonymous": "always",
+      "named": "never",
+      "asyncArrow": "always"
+    }],
+  },
+  "plugins": [
+      "react"
+  ],
+  "globals": {
+      "React": true
+  },
+  "ecmaFeatures": {
+      "jsx": true
+  },
+  "parserOptions": {
+    "sourceType": "module",
+    "ecmaVersion": 6
+  }
+}
