+++ .stylelintrc
@@ -1,0 +1,4 @@
+{
+  "ignoreFiles": ["**/node_modules/**/*.*"],
+  "extends": "stylelint-config-standard"
+}
