+++ client/drums/drum-machine.js
@@ -1,32 +1,32 @@
 const { Howl } = require('howler')
 
 var kick = new Howl({
-  src: ['./drums/kick.wav'],
+  src: ['./sound-files/kick.wav'],
   volume: 0.6,
 })
 
 var snare = new Howl({
-  src: ['./drums/snare.wav'],
+  src: ['./sound-files/snare.wav'],
   volume: 0.6,
 })
 
 var hihat = new Howl({
-  src: ['./drums/hihat_closed.wav'],
+  src: ['./sound-files/hihat_closed.wav'],
   volume: 0.6,
 })
 
 var crash = new Howl({
-  src: ['./drums/crash.wav'],
+  src: ['./sound-files/crash.wav'],
   volume: 0.6,
 })
 
 var tom_hi = new Howl({
-  src: ['./drums/tom_hi.wav'],
+  src: ['./sound-files/tom_hi.wav'],
   volume: 0.6,
 })
 
 var tom_med = new Howl({
-  src: ['./drums/tom_med.wav'],
+  src: ['./sound-files/tom_med.wav'],
   volume: 0.6,
 })
