+++ client/drums/sound-files/hihat_closed.wav
