+++ client/index.html
@@ -7,6 +7,6 @@
   <body>
     <div id="content">
     </div>
-    <script src="main.js"></script>
+    <script src="../server/public/main.js"></script>
   </body>
 </html>
