+++ client/index.js
@@ -1,13 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from '../components/app.js'
-import playDrums from '../components/drum_machine.js'
+import App from './components/app.js'
 
 ReactDOM.render(
   <App/>,
   document.querySelector('#content')
 )
-
-window.addEventListener('keydown', keyed => {
-  playDrums(keyed.keyCode)
-})
